fixup tests to make test program. please lint.
git-svn-id: file:///svn/unbound/trunk@1132 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ testcode/streamtcp.c
@@ -191,7 +191,7 @@ recv_one(int fd, int udp, ldns_buffer* buf)
 #endif
 			exit(1);
 		}
-		ldns_buffer_set_limit(buf, l);
+		ldns_buffer_set_limit(buf, (size_t)l);
 	}
 	printf("\nnext received packet\n");
 	log_buf(0, "data", buf);
